Add LifetimeManager for handling shutdown
We currently have duplicated handling of shutdown events in the Tracer and Security classes. Future work (e.g. Log submission and Telemetry) will like ly add further to this pattern.

Rather than keep duplicating this pattern, adding a LifetimeManager to register all the callbacks should simplify things somewhat
